Fix windows disk usage
