+++ utils/diskusage/disk_usage_windows.go
@@ -5,23 +5,37 @@ import (
 	"unsafe"
 )
 
-type DiskUsage struct {
+type WindowsDiskUsage struct {
 	freeBytes  int64
 	totalBytes int64
 	availBytes int64
 }
 
-// NewDiskUsages returns an object holding the disk usage of volumePath
+// Statfs returns file system statistics.
+func Statfs(volumePath string) (DiskUsage, error) {
+	diskUsage := NewWindowsDiskUsage(volumePath)
+	return DiskUsage{
+		AvailableBytes: diskUsage.Available(),
+		SizeBytes:      diskUsage.Size(),
+	}, nil
+}
+
+// NewWindowsDiskUsage returns an object holding the disk usage of volumePath
 // or nil in case of error (invalid path, etc)
-func NewDiskUsage(volumePath string) *DiskUsage {
+func NewWindowsDiskUsage(volumePath string) *WindowsDiskUsage {
 
 	h := syscall.MustLoadDLL("kernel32.dll")
 	c := h.MustFindProc("GetDiskFreeSpaceExW")
 
-	du := &DiskUsage{}
+	du := &WindowsDiskUsage{}
+
+	utf16Ptr, err := syscall.UTF16PtrFromString(volumePath)
+	if err != nil {
+		return nil
+	}
 
 	c.Call(
-		uintptr(unsafe.Pointer(syscall.StringToUTF16Ptr(volumePath))),
+		uintptr(unsafe.Pointer(utf16Ptr)),
 		uintptr(unsafe.Pointer(&du.freeBytes)),
 		uintptr(unsafe.Pointer(&du.totalBytes)),
 		uintptr(unsafe.Pointer(&du.availBytes)))
@@ -29,27 +43,12 @@ func NewDiskUsage(volumePath string) *DiskUsage {
 	return du
 }
 
-// Free returns total free bytes on file system
-func (du *DiskUsage) Free() uint64 {
-	return uint64(du.freeBytes)
-}
-
 // Available returns total available bytes on file system to an unprivileged user
-func (du *DiskUsage) Available() uint64 {
+func (du *WindowsDiskUsage) Available() uint64 {
 	return uint64(du.availBytes)
 }
 
 // Size returns total size of the file system
-func (du *DiskUsage) Size() uint64 {
+func (du *WindowsDiskUsage) Size() uint64 {
 	return uint64(du.totalBytes)
-}
-
-// Used returns total bytes used in file system
-func (du *DiskUsage) Used() uint64 {
-	return du.Size() - du.Free()
-}
-
-// Usage returns percentage of use on the file system
-func (du *DiskUsage) Usage() float32 {
-	return float32(du.Used()) / float32(du.Size())
 }
